requirements.txt
streamlit
pandas
matplotlib
fpdf
pytesseract
rapidfuzz
Pillow
pdfplumber